+++ src/main/kotlin/com/github/atm1020/tuilaunch/toolwindow/TuiLaunchToolWindowFactory.kt
@@ -16,5 +16,9 @@ class TuiLaunchToolWindowFactory : ToolWindowFactory, DumbAware {
         )
     }
 
-    override fun createToolWindowContent(project: Project, toolWindow: ToolWindow) = Unit
+    override fun createToolWindowContent(project: Project, toolWindow: ToolWindow) {
+        toolWindow.setTitleActions(
+            listOfNotNull(ActionManager.getInstance().getAction("TUILaunch.OpenPromptFile"))
+        )
+    }
 }